+++ src/app/Shield.ts
@@ -25,49 +25,49 @@ export class Shield extends Equipment {
     public hitpoints: number = 0;
     //What kind of shield is this based on?
     public shieldBase: string = "";
-    public $shieldAlly: boolean = false;
+    public _shieldAlly: boolean = false;
     //Shoddy shields take a -2 penalty to AC.
-    public $shoddy: number = 0;
+    public _shoddy: number = 0;
     get_Shoddy(creature: Creature, characterService: CharacterService) {
         //Shoddy items have a -2 penalty to AC, unless you have the Junk Tinker feat and have crafted the item yourself.
         if (this.shoddy && characterService.get_Feats("Junk Tinker")[0]?.have(creature, characterService) && this.crafted) {
-            this.$shoddy = 0;
+            this._shoddy = 0;
             return 0;
         } else if (this.shoddy) {
-            this.$shoddy = -2;
+            this._shoddy = -2;
             return -2;
         } else {
-            this.$shoddy = 0;
+            this._shoddy = 0;
             return 0;
         }
     }
     get_ShieldAlly(creature: Creature, characterService: CharacterService) {
-        this.$shieldAlly = characterService.get_Feats("Divine Ally: Shield Ally")[0]?.have(creature, characterService) && true;
-        return this.$shieldAlly;
+        this._shieldAlly = characterService.get_Feats("Divine Ally: Shield Ally")[0]?.have(creature, characterService) && true;
+        return this._shieldAlly;
     }
     get_Hardness() {
         let hardness = this.hardness;
         this.material.forEach((material: ShieldMaterial) => {
             hardness = material.hardness;
         })
-        return hardness + (this.$shieldAlly ? 2 : 0);
+        return hardness + (this._shieldAlly ? 2 : 0);
     }
     get_MaxHP() {
         let hitpoints = this.hitpoints;
         this.material.forEach((material: ShieldMaterial) => {
             hitpoints = material.hitpoints;
         })
-        return hitpoints + (this.$shieldAlly ? (Math.floor(hitpoints / 2)) : 0);
+        return hitpoints + (this._shieldAlly ? (Math.floor(hitpoints / 2)) : 0);
     }
     get_BrokenThreshold() {
         let brokenThreshold = this.brokenThreshold;
         this.material.forEach((material: ShieldMaterial) => {
             brokenThreshold = material.brokenThreshold;
         })
-        return brokenThreshold + (this.$shieldAlly ? (Math.floor(brokenThreshold / 2)) : 0);
+        return brokenThreshold + (this._shieldAlly ? (Math.floor(brokenThreshold / 2)) : 0);
     }
     get_ACBonus() {
-        return this.acbonus + this.$shoddy;
+        return this.acbonus + this._shoddy;
     }
     get_HitPoints() {
         this.damage = Math.max(Math.min(this.get_MaxHP(), this.damage), 0);