+++ src/app/activity/activity.component.ts
@@ -275,7 +275,7 @@ export class ActivityComponent implements OnInit {
                     while (!gain.effectChoices.length || gain.effectChoices.length < index - 1) {
                         gain.effectChoices.push({ condition: conditionSet.condition.name, choice: conditionSet.condition.choice });
                     }
-                    if (!conditionSet.condition.$choices.includes(gain.effectChoices?.[index]?.choice)) {
+                    if (!conditionSet.condition._choices.includes(gain.effectChoices?.[index]?.choice)) {
                         gain.effectChoices[index] = { condition: conditionSet.condition.name, choice: conditionSet.condition.choice };
                     }
                 })
@@ -304,7 +304,7 @@ export class ActivityComponent implements OnInit {
                     while (!gain.spellEffectChoices[spellCastIndex].length || gain.spellEffectChoices[spellCastIndex].length < index - 1) {
                         gain.spellEffectChoices[spellCastIndex].push({ condition: conditionSet.condition.name, choice: conditionSet.condition.choice });
                     }
-                    if (!conditionSet.condition.$choices.includes(gain.spellEffectChoices[spellCastIndex]?.[index]?.choice)) {
+                    if (!conditionSet.condition._choices.includes(gain.spellEffectChoices[spellCastIndex]?.[index]?.choice)) {
                         gain.spellEffectChoices[spellCastIndex][index] = { condition: conditionSet.condition.name, choice: conditionSet.condition.choice };
                     }
                 })