+++ src/app/itemTarget/itemTarget.component.ts
@@ -91,7 +91,7 @@ export class ItemTargetComponent implements OnInit {
                 targets.push(Object.assign(new SpellTarget(), { name: otherCreature.name || otherCreature.type, id: otherCreature.id, playerId: character.id, type: otherCreature.type, selected: false }));
             })
         }
-        if (character.partyName && !this.excluding) {
+        if (character.partyName && !this.excluding && !this.characterService.get_GMMode() && !this.characterService.get_ManualMode()) {
             //Only allow selecting other players if you are in a party.
             this.savegameService.get_Savegames().filter(savegame => savegame.partyName == character.partyName && savegame.id != character.id).forEach(savegame => {
                 targets.push(Object.assign(new SpellTarget(), { name: savegame.name || "Unnamed", id: savegame.id, playerId: savegame.id, type: "Character", selected: false }));