+++ src/app/savegame.service.ts
@@ -467,16 +467,16 @@ export class SavegameService {
                     //Delete attributes that are in the "neversave" list, if it exists.
                     if (object["neversave"] && object["neversave"].includes(key)) {
                         delete object[key];
-                        //Don't cleanup the "_className" or any attributes that are in the "save" list.
-                    } else if (!object["save"]?.includes(key) && (key != "_className") && (key.substr(0, 1) != "$")) {
+                        //Don't cleanup the "_className" or any attributes that are in the "save" list or start with "_" (which is done further down).
+                    } else if (!object["save"]?.includes(key) && (key != "_className") && (key.substr(0, 1) != "_")) {
                         //If the attribute has the same value as the default, delete it from the object.
                         if (JSON.stringify(object[key]) == JSON.stringify(blank[key])) {
                             delete object[key];
                         } else {
                             object[key] = this.clean(object[key], itemsService)
                         }
-                        //Cleanup attributes that start with $.
-                    } else if (key.substr(0, 1) == "$") {
+                        //Cleanup attributes that start with _ (except _className).
+                    } else if ((key != "_className") && key.substr(0, 1) == "_") {
                         delete object[key];
                     }
                 })