+++ Changelog.md
@@ -20,3 +20,4 @@ This log tracks only major changes, as well as changes that need characters to b
 ## 1.0.3
 
 - Rogues created before version 1.0.3 will have their "Racket" choice renamed to "Rogue's Racket" to match the correct class choice name.
+- Players are now able to send items from their inventories to other party members.